requirements.txt
flask==2.3.3
requests==2.31.0
spotipy==2.23.0
ytmusicapi==0.24.1
python-dotenv==1.0.0
flask-cors==4.0.0
gunicorn==21.2.0
python-decouple==3.8 